chore(*): move flag types to separate package
